requirements.txt
# Core Libraries
pandas
pydantic
python-dotenv

# OpenAI Integration
langchain
langchain-core
langchain-openai

# Image Processing
Pillow
